refactor: приводим всё в единую структуру
