+++ src/denoise_bot/routers/cats/original_5477554688.jpg
