+++ src/denoise_bot/routers/cats/unnamed.jpg
