+++ src/denoise_bot/routers/remove_noise_button.py
